koin + increment + Clean Architecture
