--- app/src/main/java/com/example/app4a/appViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.app4a
-
-import androidx.lifecycle.ViewModel
-
-class appViewModel : ViewModel() {
-
-}
