+++ app/src/main/java/com/example/app4a/injection/KoinModule.kt
@@ -1,13 +1,9 @@
 package com.example.app4a.injection
 
+import com.example.app4a.presentation.main.appViewModel
 import org.koin.dsl.module
 
-// Given some classes
-class Controller(val service : BusinessService)
-class BusinessService()
 
-// just declare it
 val myModule = module {
-    single { Controller(get()) }
-    single { BusinessService() }
+    factory { appViewModel() }
 }
