+++ app/src/main/java/com/example/app4a/injection/appApplication.kt
@@ -7,11 +7,8 @@ import org.koin.core.context.startKoin
 class appApplication : Application() {
     override fun onCreate(){
         super.onCreate()
-        // start Koin!
         startKoin {
-            // Android context
             androidContext(this@appApplication)
-            // modules
             modules(myModule)
         }
     }
