+++ app/src/main/java/com/example/app4a/presentation/main/MainActivity.kt
@@ -1,0 +1,25 @@
+package com.example.app4a.presentation.main
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.lifecycle.Observer
+import com.example.app4a.R
+import kotlinx.android.synthetic.main.activity_main.*
+import org.koin.android.ext.android.inject
+
+class MainActivity : AppCompatActivity() {
+
+    val mainViewModel : appViewModel by inject()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        iButton.setOnClickListener{
+            mainViewModel.onClickInc()
+        }
+        mainViewModel.livedatatest.observe(this, Observer {
+            middleText.text = it.toString()
+        })
+    }
+}
