+++ app/src/main/java/com/example/app4a/presentation/main/appViewModel.kt
@@ -1,0 +1,17 @@
+package com.example.app4a.presentation.main
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class appViewModel : ViewModel() {
+
+    val livedatatest : MutableLiveData<Int> = MutableLiveData()
+
+    init {
+        livedatatest.value = 0
+    }
+
+    fun onClickInc() {
+        livedatatest.value = (livedatatest.value ?: 0) +1 // "livedatatest?.value ?: 0" marche aussi
+    }
+}
